+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 This file is used to list changes made in each version of the Java cookbook.
 
-## Unrleased
+## Unreleased
 
 - Add aarch64 installation candidate for Corretto